split interface and implementations levels
